@@ -13,7 +13,7 @@
         "doctrine/dbal": "^3.3",
         "filament/filament": "^2.14",
         "guzzlehttp/guzzle": "^7.2",
-        "laravel/framework": "^9.21",
+        "laravel/framework": "^10.5",
         "laravel/socialite": "^5.5",
         "laravel/tinker": "^2.7",
         "maatwebsite/excel": "^3.1",
@@ -22,23 +22,23 @@
         "socialiteproviders/manager": "^4.1",
         "spatie/laravel-activitylog": "^4.5",
         "spatie/laravel-markdown": "^2.2",
-        "z3d0x/filament-logger": "^0.4.0"
+        "z3d0x/filament-logger": "^0.5"
     },
     "require-dev": {
         "barryvdh/laravel-debugbar": "^3.6",
         "barryvdh/laravel-ide-helper": "^2.12",
         "fakerphp/faker": "^1.9.1",
-        "laravel-lang/lang": "^10.7",
-        "laravel-lang/publisher": "^13.0",
+        "laravel-lang/lang": "^12.0",
+        "laravel-lang/publisher": "^14.0",
         "laravel/pint": "^1.0.0",
         "laravel/sail": "^1.0.1",
         "mockery/mockery": "^1.4.4",
-        "nunomaduro/collision": "^6.1",
-        "pestphp/pest": "^1.21",
-        "pestphp/pest-plugin-laravel": "^1.2",
-        "pestphp/pest-plugin-livewire": "^1.0",
-        "phpunit/phpunit": "^9.5.10",
-        "spatie/laravel-ignition": "^1.0"
+        "nunomaduro/collision": "^7.0",
+        "pestphp/pest": "^2.0",
+        "pestphp/pest-plugin-laravel": "^2.0",
+        "pestphp/pest-plugin-livewire": "^2.0",
+        "phpunit/phpunit": "^10.0",
+        "spatie/laravel-ignition": "^2.0"
     },
     "autoload": {
         "psr-4": {
@@ -82,7 +82,8 @@
         "preferred-install": "dist",
         "sort-packages": true,
         "allow-plugins": {
-            "pestphp/pest-plugin": true
+            "pestphp/pest-plugin": true,
+            "php-http/discovery": true
         }
     },
     "minimum-stability": "dev",
